Add XMLSerialize: explicit XML declaration (SQL/XML X078)
This adds the options INCLUDING XMLDECLARATION and EXCLUDING
XMLDECLARATION to XMLSerialize, which include or remove the
XML declaration of a given DOCUMENT or CONTENT, respectively.

SELECT
  xmlserialize(DOCUMENT val AS text INCLUDING XMLDECLARATION);

SELECT
  xmlserialize(DOCUMENT val AS text EXCLUDING XMLDECLARATION);

If not specified, the output will contain an XML declaration or
not depending on the given XML value. This patch also includes
regression tests and documentation.