Disallow dependencies from non-temporary functions to  temp objects.
We don't have an official concept of temporary functions.  (You can
make one explicitly in pg_temp, but then you have to explicitly
schema-qualify it on every call.)  However, until now we were quite
laissez-faire about whether a non-temporary function could depend on
a temporary object, such as a temp table or view.  If one did, it
would silently go away at end of session, due to the implied
DROP ... CASCADE on the session's temporary objects.  People have
complained that that's surprising, so let's fix it by disallowing
the case.  Hand-made temp functions are still allowed to reference
temp objects, so long as they're in our own session's temp schema.

This required modifying several existing regression tests that were
depending on exactly the case that's now forbidden.  Maybe that's
an indication that this is a bad idea?

We should probably make the implementation of temp-by-default
views use the same infrastructure used here, but that's for
another patch.  It's unclear whether there are any other object
classes that deserve similar treatment.

Author: Jim Jones <jim.jones@uni-muenster.de>
Reviewed-by: Tom Lane <tgl@sss.pgh.pa.us>